work in progress (recipes endpoint), refactor
